TST/REF: collect datetimelike factorize tests (#38573)
