+++ pandas/tests/indexes/period/methods/test_factorize.py
